-fixed to compile with ppc r32 and libogc 1.8.20 -had to remove wii drive light flashing during wait animation. even though i added NULL to nanosleep() it caused wiiflow lite to freeze right when the wait animation was to start. when i removed it everything works fine. besides the wii u doesn't have a disc slot light. -note - strangely going from ppc r30 to r32 wiiflow lite boot.dol grew about 24k. also somehow ext_booter.bin is now marked as changed but i don't know what's different.

+++ trunk/source/banner/Animator.cpp
@@ -232,10 +232,10 @@ float HermiteKeyHandler::GetFrame(FrameNumber frame_number) const
 
 		// curvy code from marcan, :p
 		return
-			prev->second.slope * nf * (t + powf(t, 3) - 2 * powf(t, 2)) +
-			next->second.slope * nf * (powf(t, 3) - powf(t, 2)) +
-			prev->second.value * (1 + (2 * powf(t, 3) - 3 * powf(t, 2))) +
-			next->second.value * (-2 * powf(t, 3) + 3 * powf(t, 2));
+			prev->second.slope * nf * (t + pow(t, 3) - 2 * pow(t, 2)) +
+			next->second.slope * nf * (pow(t, 3) - pow(t, 2)) +
+			prev->second.value * (1 + (2 * pow(t, 3) - 3 * pow(t, 2))) +
+			next->second.value * (-2 * pow(t, 3) + 3 * pow(t, 2));
 	}
 }
